requirements.txt
Flask==3.0.3
Flask-CORS==4.0.0
Werkzeug==3.0.3

torch==2.9.0
torchvision==0.24.0

ultralytics==8.3.10

opencv-python==4.10.0.84
Pillow==10.4.0
numpy==2.3.4
timm==1.0.9
requests==2.32.3
python-dotenv==1.0.1

psycopg2-binary==2.9.11
